fix after rebase

+++ central/usage/store/cache/cache_test.go
@@ -25,25 +25,25 @@ func TestFilterCurrent(t *testing.T) {
 
 	ids := set.NewStringSet()
 	bm := c.FilterCurrent(ids)
-	assert.Equal(t, int32(0), bm.NumNodes)
-	assert.Equal(t, int32(0), bm.NumCpuUnits)
+	assert.Equal(t, int64(0), bm.NumNodes)
+	assert.Equal(t, int64(0), bm.NumCpuUnits)
 
 	inject(c)
 
 	ids.Add("test1")
 	bm = c.FilterCurrent(ids) // removes test2 values, as not present in ids
-	assert.Equal(t, int32(1), bm.NumNodes)
-	assert.Equal(t, int32(10), bm.NumCpuUnits)
+	assert.Equal(t, int64(1), bm.NumNodes)
+	assert.Equal(t, int64(10), bm.NumCpuUnits)
 	ids.Add("test2")
 	bm = c.FilterCurrent(ids)
-	assert.Equal(t, int32(1), bm.NumNodes)
-	assert.Equal(t, int32(10), bm.NumCpuUnits)
+	assert.Equal(t, int64(1), bm.NumNodes)
+	assert.Equal(t, int64(10), bm.NumCpuUnits)
 
 	inject(c)
 
 	bm = c.FilterCurrent(ids)
-	assert.Equal(t, int32(3), bm.NumNodes)
-	assert.Equal(t, int32(30), bm.NumCpuUnits)
+	assert.Equal(t, int64(3), bm.NumNodes)
+	assert.Equal(t, int64(30), bm.NumCpuUnits)
 }
 
 func TestCutMetrics(t *testing.T) {
@@ -53,24 +53,24 @@ func TestCutMetrics(t *testing.T) {
 
 	ids := set.NewStringSet()
 	bm := c.CutMetrics(ids)
-	assert.Equal(t, int32(0), bm.NumNodes)
-	assert.Equal(t, int32(0), bm.NumCpuUnits)
+	assert.Equal(t, int64(0), bm.NumNodes)
+	assert.Equal(t, int64(0), bm.NumCpuUnits)
 
 	inject(c)
 
 	ids.Add("test1")
 	bm = c.CutMetrics(ids)
-	assert.Equal(t, int32(1), bm.NumNodes)
-	assert.Equal(t, int32(10), bm.NumCpuUnits)
+	assert.Equal(t, int64(1), bm.NumNodes)
+	assert.Equal(t, int64(10), bm.NumCpuUnits)
 
 	bm = c.CutMetrics(ids)
-	assert.Equal(t, int32(0), bm.NumNodes)
-	assert.Equal(t, int32(0), bm.NumCpuUnits)
+	assert.Equal(t, int64(0), bm.NumNodes)
+	assert.Equal(t, int64(0), bm.NumCpuUnits)
 
 	inject(c)
 
 	ids.Add("test2")
 	bm = c.CutMetrics(ids)
-	assert.Equal(t, int32(3), bm.NumNodes)
-	assert.Equal(t, int32(30), bm.NumCpuUnits)
+	assert.Equal(t, int64(3), bm.NumNodes)
+	assert.Equal(t, int64(30), bm.NumCpuUnits)
 }
